ENH: Change thick slab range to accommodate MicroCT volumes
The voxel spacing in MicroCT is much smaller than in regular medical images, so the 1mm minimum is not sensible. Reduce the minimum, the default, and the step size as well.

Related conversation (in case a change can be integrated into Slicer later): https://discourse.slicer.org/t/thick-slab-reconstruction-does-not-work-with-microct/38311

Re https://github.com/EBATINCA/Sigray3DViewer/issues/20