ENH: Move roles enum to qMRMLItemDelegate to consolidate access
Previously the qMRMLColorModel was in MRMLWidgets, but after moving it to the Colors module (so that the Color widgets have access to Terminologies widgets) the delegate did not have access to the enum. With the enum moved to the delegate now this is solved.

The PointerRole was deleted from the enum because it was not used anywhere.

Re #6975, #7593