ENH: Store terminology context list instead of one in application settings
Until now the terminology context that was last selected in the combobox was loaded by default in the terminology selector (navigator). Now a string list is stored, with the last accepted (not in the selector combobox but when accepting in the dialog) terminology context as the first entry, then the second last, and so on.

Re #6975, #7593