ENH: Return the segment ID when adding or importing from model
With the new unique segment IDs, there is no easy way to know the ID of the segment that is added or imported from model. The last segment ID could be queried, but it is not something easy nor explicit. Now these methods return the segment ID on success, and return empty string otherwise.